create new driver for OLED
